refactor(worker): block contracts as data (stage 12-2)
