+++ apps/dashboard/components/cockpit/prompt-editor/prompt-slot-fields.tsx
@@ -2,6 +2,7 @@
 
 import { useEffect, useRef, useState } from "react";
 import {
+  PROMPT_SLOT_NAME_PATTERN,
   type JsonSchema202012,
   type JsonValue,
   type PromptSlotBinding,
@@ -9,7 +10,7 @@ import {
   type WorkflowDataCatalogEntry,
   type WorkflowDataReferenceV2,
 } from "@shared/contracts";
-import { isPromptSlotBinding, PROMPT_SLOT_NAME_PATTERN } from "@shared/prompts";
+import { isPromptSlotBinding } from "@shared/prompts";
 import { JsonSchemaEditor } from "@/components/cockpit/flow-editor/json-schema-editor";
 import type { JsonSchemaEditorValidationState } from "@/components/cockpit/flow-editor/json-schema-editor";
