+++ apps/worker/src/mcp/tools/blocks.ts
@@ -2,10 +2,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
 import type { WorkflowBlockContract } from "@shared/contracts";
 
-import {
-  buildWorkflowBlockRegistry,
-  workflowBlockRegistryContextFromEnv,
-} from "../../services/mcp/app-dependencies.js";
+import { currentBlockContracts } from "../../services/workflow-definitions/block-contracts.js";
 import { McpPublicError, type McpToolDependencies } from "../contracts.js";
 import { executeMcpRead } from "../execute-tool.js";
 import { registerCatalogTool } from "../tool-catalog.js";
@@ -15,14 +12,13 @@ type BlocksListData = {
 };
 
 // Environment-derived (which agent, VCS and messaging providers this deployment
-// has configured) and cheap to recompute -- the same function
-// buildWorkflowEditorOptions calls on every dashboard load (workflow-definition/
-// models.ts:120) -- so it is read fresh per call rather than cached. A provider
-// that comes online mid-process (a rotated token, a newly configured Slack
-// channel) is then visible on the very next call instead of waiting for a
-// restart.
+// has configured) and cheap to recompute -- the same block data the dashboard's
+// editor read binds on every load -- so it is read fresh per call rather than
+// cached. A provider that comes online mid-process (a rotated token, a newly
+// configured Slack channel) is then visible on the very next call instead of
+// waiting for a restart.
 function buildRegistry(): Record<string, WorkflowBlockContract> {
-  return buildWorkflowBlockRegistry(workflowBlockRegistryContextFromEnv());
+  return currentBlockContracts().blockRegistry();
 }
 
 export function registerBlockTools(server: McpServer, deps: McpToolDependencies): void {
