+++ apps/worker/src/mcp/tools/workflow-authoring.ts
@@ -15,9 +15,9 @@ import {
   declaresRetiredSchema,
   logger,
   validateWorkflowDefinitionCandidate,
-  workflowBlockRegistryContextFromEnv,
 } from "../../services/mcp/app-dependencies.js";
 import { isRepoAllowed } from "../../services/dispatch/repo-allowlist.js";
+import { currentBlockContracts } from "../../services/workflow-definitions/block-contracts.js";
 import {
   runnableDefinitionOf,
   WorkflowDefinitionStoreError,
@@ -488,9 +488,12 @@ export function registerWorkflowAuthoringTools(
           // authority on a legal graph is applied, and the store then parses the
           // same schema again before it stores anything.
           //
+          const contracts = currentBlockContracts();
           const candidate = validateWorkflowDefinitionCandidate(
             input.definition,
-            workflowBlockRegistryContextFromEnv(),
+            contracts.resolveContract,
+            contracts.blockParamsSchemas,
+            contracts.configuredVcsProviders,
           );
           if (!candidate.parsed) {
             throw refusal(
