+++ apps/worker/src/workflow-definition/prompt-authoring.test.ts
@@ -5,7 +5,8 @@ import type {
   WorkflowDefinitionV2,
   WorkflowDefinitionV2Node,
 } from "@shared/contracts";
-import type { WorkflowBlockRegistryContext } from "./block-registry.js";
+import type { WorkflowBlockRegistryContext } from "../engine/definition/block-contract-resolver.js";
+import { testBlockContractResolver } from "../test-support/block-contracts.js";
 import {
   resolveNodePromptAuthoring,
   validateWorkflowPromptAuthoringIssuesWithLoader,
@@ -21,6 +22,8 @@ const registryContext: WorkflowBlockRegistryContext = {
   arthurConfigured: true,
   webhookTriggerConfigured: true,
 };
+
+const resolveContract = testBlockContractResolver(registryContext);
 
 const slot = (
   schema: PromptSlotDefinition["schema"] = { type: "string" },
@@ -205,7 +208,7 @@ describe("v2 prompt authoring validation", () => {
 
     const issues = await validateWorkflowPromptAuthoringIssuesWithLoader(
       definition,
-      registryContext,
+      resolveContract,
       async () => ({
         promptId: 1,
         promptName: "Implementation",
@@ -339,7 +342,7 @@ describe("v2 prompt authoring validation", () => {
     };
     const issues = await validateWorkflowPromptAuthoringIssuesWithLoader(
       definition,
-      registryContext,
+      resolveContract,
       async () => ({
         promptId: 1,
         promptName: "Shared",
@@ -431,7 +434,7 @@ describe("v2 prompt authoring validation", () => {
 
     const issues = await validateWorkflowPromptAuthoringIssuesWithLoader(
       definition,
-      registryContext,
+      resolveContract,
       vi.fn(),
     );
 
@@ -504,7 +507,7 @@ describe("v2 prompt authoring validation", () => {
 
     const issues = await validateWorkflowPromptAuthoringIssuesWithLoader(
       definition,
-      registryContext,
+      resolveContract,
       vi.fn(),
     );
 
@@ -559,7 +562,7 @@ describe("v2 prompt authoring validation", () => {
 
     const issues = await validateWorkflowPromptAuthoringIssuesWithLoader(
       definition,
-      registryContext,
+      resolveContract,
       vi.fn(),
     );
