+++ apps/worker/src/workflow-definition/schema.ts
@@ -12,13 +12,14 @@ import type {
   WorkflowDefinitionV2,
   WorkflowDefinitionV2Node,
   WorkflowDefinitionValidationIssue,
+  WorkflowBlockContractResolver,
   WorkflowParamValue,
+  VcsProviderKind,
 } from "@shared/contracts";
 import {
   BLOCK_PARAM_KEYS,
   BLOCK_TYPE_SPECS,
   FAILURE_PORT,
-  WEBHOOK_AUTH_SCHEMES,
   isHarnessProfileReference,
   isSafeWorkflowInputName,
   isTriggerBlockType,
@@ -27,34 +28,14 @@ import {
   evaluateWorkflowValueCompatibility,
 } from "@shared/contracts";
 import { resolveBuiltinHarnessProfile } from "@shared/harness";
-import { PROMPT_SLOT_NAME_PATTERN } from "@shared/prompts";
-import {
-  prepareWorkspaceParams,
-  finalizeWorkspaceParams,
-  fixAgentParams,
-  genericAgentParams,
-  callLlmParams,
-  fetchPrContextParams,
-  investigateParams,
-  runChecksParams,
-  postTicketCommentParams,
-  postPrCommentParams,
-  humanQuestionParams,
-  arthurInjectionCheckParams,
-  leakReviewParams,
-  sendPlanApprovalParams,
-} from "../engine/definition/params.generated.js";
-import { repositoryScriptGroupNameSchema } from "../engine/blocks/run-checks/manifest.js";
 import {
   MINIMUM_PERIOD_MS,
   parseSchedule,
   violatesMinimumPeriod,
 } from "../engine/definition/schedule-occurrence.js";
 import {
-  resolveWorkflowBlockContract,
   workflowBlockDeploymentDefinitionIssues,
   workflowRepositoryScopeIssues,
-  type WorkflowBlockRegistryContext,
 } from "./block-registry.js";
 import {
   analyzeWorkflowV2Bindings,
@@ -62,6 +43,14 @@ import {
 } from "./available-values.js";
 import { validateTransformDefinition } from "./transform.js";
 import { validateWorkflowV2WorkspaceAccessIssues } from "./workspace-access.js";
+
+/**
+ * The per-type block parameter parsers these rules validate a definition
+ * against. Declared where the parameter is taken, not where the map is
+ * composed (`engine/definition/block-params-schemas.ts`), so these rules keep
+ * no import back to the block modules the map pulls in.
+ */
+export type WorkflowBlockParamsSchemas = Record<WorkflowBlockType, z.ZodTypeAny>;
 const nodeId = z.string().trim().min(1);
 const coordinate = z.number().finite();
 const bindingInputName = z.custom<string>(
@@ -69,36 +58,18 @@ const bindingInputName = z.custom<string>(
   { message: "Input name contains an empty or unsafe path segment." },
 );
 
-const emptyParams = z.object({}).strict();
-const agentParams = z
-  .object({
-    model: z.string().trim().max(200).regex(/^[A-Za-z0-9._:/-]+$/).optional(),
-    provider: z.enum(["claude", "codex"]).optional(),
-    prompt: z.string().trim().min(1).max(50000).optional(),
-  })
-  .strict();
-
 const vcsProviders = z.enum(["github", "gitlab"]);
-const vcsProviderSelection = z.array(vcsProviders).min(1);
-const reviewStates = z.enum(["changes_requested", "commented"]);
-const prTriggerScope = z.enum(["workflow_owned", "any"]);
-
-/** Optional per-node start budget. Both keys optional: absent means unlimited.
- * rateLimitWindow is the fixed UTC window rateLimitMax applies to. */
-const triggerRateLimitParams = {
-  rateLimitMax: z.number().int().min(1).optional(),
-  rateLimitWindow: z.enum(["minute", "hour", "day", "month"]).optional(),
-};
+export const vcsProviderSelection = z.array(vcsProviders).min(1);
 
 // Sized far above any hand-drawn workflow (the built-in default is 8 blocks/7
 // connections) but low enough to bound validateWorkflowGraph, whose dominator
 // fixpoint is O(N^2*E) and copies the node universe per node.
 //
 // Exported so the MCP tool catalog's own size gate can be pinned against them
-// (mcp/tool-catalog.test.ts): the catalog restates the numbers rather than importing
-// them, because it is loaded on the transport path and this module pulls in every
-// block, and a restated number that drifts below these would leave an agent able to
-// read a graph it can never save back.
+// (mcp/tool-catalog.test.ts): the catalog restates the numbers rather than
+// importing them, because it is loaded on the transport path and must stay out
+// of this module's graph, and a restated number that drifts below these would
+// leave an agent able to read a graph it can never save back.
 export const MAX_NODES = 200;
 export const MAX_EDGES = 400;
 const executionBudgetsSchema = z
@@ -194,7 +165,7 @@ export function isWorkflowDataReferenceV2(
   );
 }
 
-const workflowInputBindingV2Schema = z.discriminatedUnion(
+export const workflowInputBindingV2Schema = z.discriminatedUnion(
   "kind",
   [
     z
@@ -260,7 +231,7 @@ const transformBuildObjectFieldSchema = z
     value: transformBuildObjectValueSchema,
   })
   .strict();
-const transformConfigurationSchema: z.ZodType<TransformConfiguration> = z.discriminatedUnion(
+export const transformConfigurationSchema: z.ZodType<TransformConfiguration> = z.discriminatedUnion(
   "operation",
   [
     z.object({ operation: z.literal("format_text"), template: z.string() }).strict(),
@@ -299,199 +270,9 @@ const transformConfigurationSchema: z.ZodType<TransformConfiguration> = z.discri
   ],
 );
 
-const v2TriggerPrCreatedConfiguration = z
-  .object({
-    providers: vcsProviderSelection.default(["github", "gitlab"]),
-    scope: prTriggerScope.default("workflow_owned"),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-const v2TriggerPrReadyConfiguration = z
-  .object({
-    providers: vcsProviderSelection.default(["github", "gitlab"]),
-    scope: prTriggerScope.default("any"),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-const v2TriggerPrUpdatedConfiguration = v2TriggerPrReadyConfiguration;
-const v2TriggerPrChecksFailedConfiguration = z
-  .object({
-    providers: vcsProviderSelection.default(["github", "gitlab"]),
-    scope: prTriggerScope.default("workflow_owned"),
-    checkNames: z.array(z.string().trim().min(1).max(255)).max(100).default([]),
-    ignoreCheckNames: z.array(z.string().trim().min(1).max(255)).max(100).default([]),
-    githubAppSlugs: z
-      .array(z.string().trim().min(1).max(100))
-      .min(1)
-      .max(20)
-      .default(["github-actions"]),
-    gitlabPipelineSources: z
-      .array(z.string().trim().min(1).max(100))
-      .min(1)
-      .max(20)
-      .default(["merge_request_event"]),
-    maxFixAttemptsPerPr: z.number().int().min(1).max(10).default(2),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-const v2TriggerPrReviewConfiguration = z
-  .object({
-    providers: vcsProviderSelection.default(["github"]),
-    on: z.array(reviewStates).min(1).default(["changes_requested"]),
-    scope: prTriggerScope.default("workflow_owned"),
-    maxRunsPerPr: z.number().int().min(1).max(30).default(10),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-const v2TriggerPrMergedConfiguration = z
-  .object({
-    providers: vcsProviderSelection.default(["github", "gitlab"]),
-    scope: prTriggerScope.default("workflow_owned"),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-/** Dot-path into the delivered JSON body ("ticket.subject"). Reuses the shared
- * segment rule (`[A-Za-z0-9_-]+` per segment, no prototype-mutating names) so a
- * mapping authored here cannot traverse anywhere a binding could not. */
-const webhookPayloadPath = z
-  .string()
-  .trim()
-  .min(1)
-  .max(200)
-  .refine(isSafeWorkflowInputName, {
-    message: "Payload path contains an empty or unsafe segment.",
-  });
-/** Every key is optional: the block registry supplies the mapping defaults, and
- * the endpoint row carries the auth scheme, re-synced from this config on every
- * deploy (like any other block parameter). */
-const v2TriggerWebhookConfiguration = z
-  .object({
-    provider: z.enum(["zendesk", "sentry"]).optional(),
-    sourceIdPath: webhookPayloadPath.optional(),
-    sourceUrlPath: webhookPayloadPath.optional(),
-    customerContextPath: webhookPayloadPath.optional(),
-    authScheme: z.enum(WEBHOOK_AUTH_SCHEMES).optional(),
-    headerName: z
-      .string()
-      .trim()
-      .min(1)
-      .max(100)
-      .regex(
-        /^[A-Za-z0-9!#$%&'*+.^_`|~-]+$/,
-        "Header name must be a valid HTTP header token.",
-      )
-      .optional(),
-    requireTimestamp: z.boolean().optional(),
-    timestampHeader: z
-      .string()
-      .trim()
-      .min(1)
-      .max(100)
-      .regex(
-        /^[A-Za-z0-9!#$%&'*+.^_`|~-]+$/,
-        "Header name must be a valid HTTP header token.",
-      )
-      .optional(),
-    // Ceiling kept tight (15 minutes) so replay protection cannot be widened into
-    // a multi-hour, two-sided replay window. The default stays 300 seconds.
-    timestampToleranceSeconds: z.number().int().min(30).max(900).optional(),
-    subjectPath: webhookPayloadPath.optional(),
-    mapSubject: webhookPayloadPath.optional(),
-    mapDescription: webhookPayloadPath.optional(),
-    mapRequester: webhookPayloadPath.optional(),
-    mapPriority: webhookPayloadPath.optional(),
-    ...triggerRateLimitParams,
-  })
-  .strict()
-  // Replay protection folds the timestamp into the HMAC signed message, so it is
-  // meaningless for shared_token (a constant header has nothing to sign). Reject
-  // the combination instead of silently no-opping into a false sense of safety.
-  .superRefine((config, ctx) => {
-    if (config.requireTimestamp === true && config.authScheme === "shared_token") {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        path: ["requireTimestamp"],
-        message: "Replay protection requires the HMAC SHA-256 scheme.",
-      });
-    }
-  });
-/**
- * Smallest catch-up window an author may configure.
- *
- * The dial reads like "how stale a run may be", but what it actually buys is
- * "how many consecutive missed ticks I tolerate", because the scheduler
- * evaluates once a minute. At 1 minute, measured, a single two-minute stall of
- * the tick loses the run outright, and so does a steady 75 second delay; at 60
- * the same stall costs nothing. An author tightening this to avoid stale work
- * would instead be handing every hiccup of the platform cron a silently
- * swallowed run, so the floor is set where one lost tick is still survivable.
- */
-const MIN_CATCH_UP_GRACE_MINUTES = 5;
-
-/** Cron syntax is checked by the deployment validator, not by this schema.
- * Empty cron/taskTitle/taskDescription stay legal at this level so a
- * partially configured draft still saves; deployment separately refuses to
- * publish an incomplete one. */
-const v2TriggerScheduleConfiguration = z
-  .object({
-    cron: z.string().default(""),
-    timezone: z.string().default("UTC"),
-    overlapPolicy: z.enum(["skip", "queue", "allow"]).default("skip"),
-    /** Floor of 5, see MIN_CATCH_UP_GRACE_MINUTES. */
-    catchUpGraceMinutes: z
-      .number()
-      .int()
-      .min(
-        MIN_CATCH_UP_GRACE_MINUTES,
-        `catchUpGraceMinutes must be at least ${MIN_CATCH_UP_GRACE_MINUTES} minutes: the scheduler evaluates once a minute, so a smaller tolerance means a single missed tick silently loses the run.`,
-      )
-      .default(60),
-    taskTitle: z.string().default(""),
-    taskDescription: z.string().default(""),
-    ...triggerRateLimitParams,
-  })
-  .strict();
-/** Accepted and ignored. The repair loop maxFixCycles bounded is gone, but
- *  every definition deployed against it still carries the key, and a strict
- *  schema that drops a key stops those definitions from loading at all. The
- *  bound stays exactly as authored so nothing that used to be invalid becomes
- *  valid on the way past. */
-const v2RunPrePrChecksConfiguration = z
-  .object({ maxFixCycles: z.number().int().min(0).max(5).optional() })
-  .strict();
-/** run_scripts selects groups by name and nothing else. At least one: a node
- *  that runs no group verifies nothing while still reporting an outcome. */
-const v2RunScriptsConfiguration = z
-  .object({ groups: z.array(repositoryScriptGroupNameSchema).min(1) })
-  .strict();
-const v2OpenPrConfiguration = z
-  .object({
-    title: z.string().optional(),
-    body: z.string().optional(),
-  })
-  .strict();
-const v2UpdateTicketStatusConfiguration = z
-  .object({ target: z.string().trim().min(1).max(200) })
-  .strict();
-const v2SendSlackMessageConfiguration = z
-  .object({
-    message: z.string().trim().max(2000).optional(),
-    sendOn: z.enum(["pr_ready", "always"]).optional(),
-  })
-  .strict();
-const v2CreatePrCheckConfiguration = z
-  .object({
-    checkName: z.string().trim().min(1).max(200),
-  })
-  .strict();
-const v2CompletePrCheckConfiguration = z
-  .object({
-    conclusion: z.enum(["success", "failure", "neutral"]),
-    details: z.string().max(10_000).optional(),
-    refreshHead: z.boolean().optional(),
-  })
-  .strict();
-const v2LoopConfiguration = z
+/** Loop carries are named, typed values the next attempt reads, so the shape is
+ *  read both by the params schema map and by the loop port rules below. */
+export const v2LoopConfiguration = z
   .object({
     maxAttempts: z.number().int().min(1).max(20),
     onExhaust: z.enum(["fail", "human", "continue"]),
@@ -509,88 +290,6 @@ const v2LoopConfiguration = z
       .optional(),
   })
   .strict();
-const v2TerminateConfiguration = z
-  .object({
-    terminalStatus: z.enum([
-      "waiting_for_human",
-      "failed",
-      "skipped",
-      "done",
-    ]),
-    postComment: z.string().trim().min(1).max(2000).optional(),
-  })
-  .strict();
-const jsonSchemaDialect202012 = z
-  .literal("https://json-schema.org/draft/2020-12/schema")
-  .optional();
-const v2PromptSlotBindings = z
-  .record(
-    z.string().regex(PROMPT_SLOT_NAME_PATTERN),
-    workflowInputBindingV2Schema,
-  )
-  .optional();
-const harnessProfileReferenceSchema = z
-  .object({
-    profileId: z.string().trim().min(1).max(200),
-    version: z.number().int().positive(),
-  })
-  .strict()
-  .optional();
-const v2PromptAuthoringConfiguration = {
-  harnessProfile: harnessProfileReferenceSchema,
-  promptSlotBindings: v2PromptSlotBindings,
-};
-
-/** The v2 runtime consumes the same code-owned configuration surface as the
- * corresponding v1 executor. Transform and Branch intentionally use their own
- * typed configuration validators below. */
-const v2ConfigurationSchemas = {
-  trigger_ticket_ai: z.object(triggerRateLimitParams).strict(),
-  trigger_plan_approved: emptyParams,
-  trigger_pr_created: v2TriggerPrCreatedConfiguration,
-  trigger_pr_ready: v2TriggerPrReadyConfiguration,
-  trigger_pr_updated: v2TriggerPrUpdatedConfiguration,
-  trigger_pr_checks_failed: v2TriggerPrChecksFailedConfiguration,
-  trigger_pr_review: v2TriggerPrReviewConfiguration,
-  trigger_pr_merged: v2TriggerPrMergedConfiguration,
-  trigger_webhook: v2TriggerWebhookConfiguration,
-  trigger_schedule: v2TriggerScheduleConfiguration,
-  planning_agent: agentParams.extend(v2PromptAuthoringConfiguration),
-  implementation_agent: agentParams.extend(v2PromptAuthoringConfiguration),
-  review_agent: agentParams.extend(v2PromptAuthoringConfiguration),
-  fix_agent: fixAgentParams.extend(v2PromptAuthoringConfiguration),
-  generic_agent: genericAgentParams.extend({
-    outputSchemaDialect: jsonSchemaDialect202012,
-    ...v2PromptAuthoringConfiguration,
-  }),
-  prepare_workspace: prepareWorkspaceParams,
-  finalize_workspace: finalizeWorkspaceParams,
-  run_pre_pr_checks: v2RunPrePrChecksConfiguration,
-  run_checks: runChecksParams,
-  run_scripts: v2RunScriptsConfiguration,
-  call_llm: callLlmParams.extend({
-    outputSchemaDialect: jsonSchemaDialect202012,
-  }),
-  fetch_pr_context: fetchPrContextParams,
-  investigate: investigateParams,
-  open_pr: v2OpenPrConfiguration,
-  update_ticket_status: v2UpdateTicketStatusConfiguration,
-  post_ticket_comment: postTicketCommentParams,
-  post_pr_comment: postPrCommentParams,
-  create_pr_check: v2CreatePrCheckConfiguration,
-  complete_pr_check: v2CompletePrCheckConfiguration,
-  post_pr_review: emptyParams,
-  send_slack_message: v2SendSlackMessageConfiguration,
-  send_plan_approval: sendPlanApprovalParams,
-  human_question: humanQuestionParams,
-  arthur_injection_check: arthurInjectionCheckParams,
-  leak_review: leakReviewParams,
-  loop: v2LoopConfiguration,
-  terminate: v2TerminateConfiguration,
-} satisfies Record<
-  Exclude<WorkflowBlockType, "branch" | "transform">,
-  z.ZodTypeAny
->;
 
 const v2BranchConditionSchema = z
   .object({
@@ -611,7 +310,7 @@ const v2BranchConditionSchema = z
     ignoreCase: z.boolean().optional(),
   })
   .strict();
-const v2BranchConfigurationSchema = z
+export const v2BranchConfigurationSchema = z
   .object({
     combinator: z.enum(["all", "any"]),
     conditions: z.array(v2BranchConditionSchema).max(100),
@@ -876,6 +575,7 @@ function invalidConfigurationIssue(
 
 function validateWorkflowV2ConfigurationIssues(
   def: WorkflowDefinitionV2,
+  blockParamsSchemas: WorkflowBlockParamsSchemas,
 ): WorkflowDefinitionValidationIssue[] {
   const issues: WorkflowDefinitionValidationIssue[] = [];
   for (const [nodeIndex, node] of def.nodes.entries()) {
@@ -906,11 +606,7 @@ function validateWorkflowV2ConfigurationIssues(
       );
     }
 
-    const schema =
-      node.type === "branch"
-        ? v2BranchConfigurationSchema
-        : v2ConfigurationSchemas[node.type];
-    const parsed = schema.safeParse(node.configuration);
+    const parsed = blockParamsSchemas[node.type].safeParse(node.configuration);
     const profileReference = node.configuration.harnessProfile;
     if (
       parsed.success &&
@@ -949,11 +645,14 @@ function validateWorkflowV2ConfigurationIssues(
 
 function v2ConfigurationParams(
   node: WorkflowDefinitionV2Node,
+  blockParamsSchemas: WorkflowBlockParamsSchemas,
 ): Record<string, WorkflowParamValue> {
+  // Branch and Transform keep their configuration out of params: their typed
+  // shapes are operations, never executor params.
   const parsedConfiguration =
     node.type === "branch" || node.type === "transform"
       ? null
-      : v2ConfigurationSchemas[node.type].safeParse(node.configuration);
+      : blockParamsSchemas[node.type].safeParse(node.configuration);
   const configuration =
     parsedConfiguration?.success === true
       ? (parsedConfiguration.data as Record<string, unknown>)
@@ -991,12 +690,13 @@ function v2ConfigurationParams(
 
 function validateWorkflowV2BlockDeploymentIssues(
   def: WorkflowDefinitionV2,
-  registryContext: WorkflowBlockRegistryContext,
+  resolveContract: WorkflowBlockContractResolver,
+  blockParamsSchemas: WorkflowBlockParamsSchemas,
   options: { checkEnvironmentAvailability?: boolean },
 ): WorkflowDefinitionValidationIssue[] {
   const issues: WorkflowDefinitionValidationIssue[] = [];
   for (const [nodeIndex, node] of def.nodes.entries()) {
-    const params = v2ConfigurationParams(node);
+    const params = v2ConfigurationParams(node, blockParamsSchemas);
     if (
       node.type === "trigger_schedule" &&
       (typeof params.cron !== "string" || params.cron.trim() === "")
@@ -1138,11 +838,7 @@ function validateWorkflowV2BlockDeploymentIssues(
         })),
       );
     } else if (options.checkEnvironmentAvailability !== false) {
-      const availability = resolveWorkflowBlockContract(
-        node.type,
-        params,
-        registryContext,
-      ).availability;
+      const availability = resolveContract(node.type, params).availability;
       if (!availability.available) {
         issues.push(
           deploymentIssue(
@@ -1845,31 +1541,40 @@ function validateWorkflowGraphV2Issues(
  * operator can keep editing a structurally sound but incomplete graph. */
 export function validateWorkflowDefinitionForDeployment(
   def: WorkflowDefinition,
-  registryContext: WorkflowBlockRegistryContext,
+  resolveContract: WorkflowBlockContractResolver,
+  blockParamsSchemas: WorkflowBlockParamsSchemas,
+  configuredVcsProviders: readonly VcsProviderKind[],
   options: {
     checkEnvironmentAvailability?: boolean;
   } = {},
 ): string[] {
-  return validateWorkflowDefinitionIssuesForDeployment(def, registryContext, options).map(
-    ({ message }) => message,
-  );
+  return validateWorkflowDefinitionIssuesForDeployment(
+    def,
+    resolveContract,
+    blockParamsSchemas,
+    configuredVcsProviders,
+    options,
+  ).map(({ message }) => message);
 }
 
 export function validateWorkflowDefinitionIssuesForDeployment(
   def: WorkflowDefinition,
-  registryContext: WorkflowBlockRegistryContext,
+  resolveContract: WorkflowBlockContractResolver,
+  blockParamsSchemas: WorkflowBlockParamsSchemas,
+  configuredVcsProviders: readonly VcsProviderKind[],
   options: {
     checkEnvironmentAvailability?: boolean;
   } = {},
 ): WorkflowDefinitionValidationIssue[] {
-  const bindingAnalysis = analyzeWorkflowV2Bindings(def, registryContext);
-  const catalogAnalysis = analyzeWorkflowV2Catalog(def, registryContext);
+  const bindingAnalysis = analyzeWorkflowV2Bindings(def, resolveContract);
+  const catalogAnalysis = analyzeWorkflowV2Catalog(def, resolveContract);
   const issues = dedupeDeploymentIssues([
     ...validateWorkflowGraphV2Issues(def),
-    ...validateWorkflowV2ConfigurationIssues(def),
+    ...validateWorkflowV2ConfigurationIssues(def, blockParamsSchemas),
     ...validateWorkflowV2BlockDeploymentIssues(
       def,
-      registryContext,
+      resolveContract,
+      blockParamsSchemas,
       options,
     ),
     ...bindingAnalysis.issues,
@@ -1882,7 +1587,7 @@ export function validateWorkflowDefinitionIssuesForDeployment(
       catalogAnalysis.catalogByNode,
     ),
     ...validateWorkflowV2WorkspaceAccessIssues(def),
-    ...repositoryScopePinIssues(def, registryContext, options),
+    ...repositoryScopePinIssues(def, configuredVcsProviders, options),
   ]);
   return issues;
 }
@@ -1894,12 +1599,12 @@ export function validateWorkflowDefinitionIssuesForDeployment(
  */
 function repositoryScopePinIssues(
   def: WorkflowDefinition,
-  registryContext: WorkflowBlockRegistryContext,
+  configuredVcsProviders: readonly VcsProviderKind[],
   options: { checkEnvironmentAvailability?: boolean },
 ): WorkflowDefinitionValidationIssue[] {
   return workflowRepositoryScopeIssues(
     def.repositoryScope,
-    registryContext,
+    configuredVcsProviders,
     options,
   ).map((message) => deploymentIssue(message, null, "/repositoryScope"));
 }
